+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changelog of threedi-modelchecker
 ------------------
 
 - Ignore tiny floating-point deviations in RasterGridSizeCheck (check 798).
+- Add check 327 to make sure vegetation drag is only used if the friction type is Chezy.
 
 
 2.2.2 (2023-05-17)